pyproject.toml
[project]
name = "orgschema-framework"
version = "0.1.0"
description = "Schema validator for Organizational Schema Theory reverse-design TDD specifications"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
orgschema-validate = "orgschema_framework.validate:main"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]

[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["import-untyped"]
